Remapper Modularization
-split cc_remapper out
-changed capitalization
-did not fix _METRONOME_VELOCITY globals
-reduced sleep pauses in cc_remapper
